requirement.txt
google-genai
python-dotenv
fastapi
uvicorn
streamlit
langchain
langchain-google-genai
pydantic
chromadb